Test the significance of dose-mortality response differences

This function is used when comparing at least two strains. It tests whether the mortality-dose regressions are similar for different strains, using a likelihood ratio test (LRT). If there are more than two strains, it also computes pairwise tests, using sequential Bonferroni correction (Hommel, 1988) to account for multiple testing.

Usage

model.signif(data)

Arguments

data

a data frame of probit transformed mortality data using the function probit.trans

Value

a list with model outputs: a chi-square test if there are only two strains or if there are more than two strains, first an overall model assessment (i.e. one strain vs. all) and given overall model is significant, then a bonferroni test of significance from a pariwise model comparison.

Details

A global LRT test assesses a strain’s effect, by comparing two models, one with and one without this effect (i.e. comparing a model with several strains to a model where all the data originate from a single strain). If there are more than two strains, pairwise tests are computed, and p-values of significance are assessed using sequential Bonferroni correction (Hommel, 1988) to account for multiple testing.

Warning: We strongly encourage users to not use this function when the dose-mortality response for at least one strain significantly deviates from linearity (see resist.ratio() function for more details): in such cases the test cannot be interpreted.

Author

Pascal Milesi, Piyal Karunarathne, Pierrick Labbé

Examples

data(bioassay)
transd<-probit.trans(bioassay$assay2)
data<-transd$tr.data
model.signif(data)
#> Analysis of Deviance Table
#> 
#> Model 1: mortality ~ log10(data$dose)
#> Model 2: mortality ~ log10(data$dose) * data$strain
#>   Resid. Df Resid. Dev Df Deviance  Pr(>Chi)    
#> 1        32     936.46                          
#> 2        28      70.69  4   865.77 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> complete model is significant against a NULL model 
#>  continueing to pair-wise comparison
#> $model
#>   strain1 strain2 model.pval  deviance df bonferroni
#> 1    DZOU   DZOU2    0.26402   5.52848  2    non-sig
#> 2    DZOU KIS-ref          0 714.37345  2        sig
#> 3   DZOU2 KIS-ref          0 603.93126  2        sig
#>